Oprava chyby způsobující freeze pro výsledky >4000000000^2
Změna datového typu `size_t` při převodu doubleToString() na
`double`.

Author:    David Krejčí
Co-Author: Jan Kalina

+++ src/CalmCatCalc/core/cat_calc_core.cpp
@@ -372,7 +372,7 @@ vector<string> postfix(vector<string> parsedExpression){
 
 string doubleToString(double x){
     size_t numLen = 0;
-    size_t i = 10;
+    double i = 10;
     while(x/i > 1){
         numLen++;
         i = i*10;
